+++ packages/pangea-multipass-llama-index/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added 
+
+- GitLabReader and GitLabProcessor
+
 ### Fixed
 
 - `pydantic` error of `node_processor` default value.